+++ it/server/src/test/java/com/walmartlabs/concord/it/server/DynamicFormIT.java
@@ -1,0 +1,85 @@
+package com.walmartlabs.concord.it.server;
+
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2019 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import com.walmartlabs.concord.client.*;
+import org.junit.Test;
+
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import static com.walmartlabs.concord.it.common.ITUtils.archive;
+import static com.walmartlabs.concord.it.common.ServerClient.*;
+import static org.junit.Assert.*;
+import static org.junit.Assert.assertTrue;
+
+public class DynamicFormIT extends AbstractServerIT {
+
+    @Test(timeout = DEFAULT_TEST_TIMEOUT)
+    public void testWithGroovy() throws Exception {
+        String firstName = "firstName_" + randomString();
+        String lastName = "lastName_" + randomString();
+
+        // ---
+
+        byte[] payload = archive(ProcessIT.class.getResource("dynamicFormWithGroovy").toURI());
+
+        ProcessApi processApi = new ProcessApi(getApiClient());
+        StartProcessResponse spr = start(payload);
+        assertNotNull(spr.getInstanceId());
+
+        // ---
+
+        ProcessEntry pir = waitForStatus(processApi, spr.getInstanceId(), ProcessEntry.StatusEnum.SUSPENDED);
+
+        // ---
+
+        ProcessFormsApi formsApi = new ProcessFormsApi(getApiClient());
+
+        List<FormListEntry> forms = formsApi.list(spr.getInstanceId());
+        assertEquals(1, forms.size());
+
+        // ---
+
+        FormListEntry f0 = forms.get(0);
+        assertFalse(f0.isCustom());
+
+        String formName = f0.getName();
+
+        Map<String, Object> data = new HashMap<>();
+        data.put("firstName", firstName);
+        data.put("lastName", lastName);
+        FormSubmitResponse fsr = formsApi.submit(pir.getInstanceId(), formName, data);
+        assertTrue(fsr.isOk());
+
+        // ---
+
+        pir = waitForCompletion(processApi, spr.getInstanceId());
+        assertEquals(ProcessEntry.StatusEnum.FINISHED, pir.getStatus());
+
+        // --
+
+        byte[] ab = getLog(pir.getLogFileName());
+        assertLog(".*firstName.*", ab);
+    }
+}