+++ runner/src/main/java/com/walmartlabs/concord/runner/engine/ConcordExecutionContextFactory.java
@@ -21,13 +21,18 @@ package com.walmartlabs.concord.runner.engine;
  */
 
 import com.walmartlabs.concord.project.InternalConstants;
+import com.walmartlabs.concord.sdk.Constants;
 import com.walmartlabs.concord.sdk.Context;
 import io.takari.bpm.api.ExecutionContext;
 import io.takari.bpm.api.ExecutionContextFactory;
+import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.api.Variables;
 import io.takari.bpm.context.DefaultExecutionContextFactory;
 import io.takari.bpm.context.ExecutionContextImpl;
 import io.takari.bpm.el.ExpressionManager;
+import io.takari.bpm.form.FormService;
+import io.takari.bpm.model.form.FormDefinition;
+import io.takari.bpm.model.form.FormField;
 
 import java.util.*;
 
@@ -35,20 +40,22 @@ public class ConcordExecutionContextFactory implements ExecutionContextFactory<C
 
     private final ExpressionManager expressionManager;
     private final ProtectedVarContext protectedVarContext;
+    private final FormService formService;
 
-    public ConcordExecutionContextFactory(ExpressionManager expressionManager, ProtectedVarContext protectedVarContext) {
+    public ConcordExecutionContextFactory(ExpressionManager expressionManager, ProtectedVarContext protectedVarContext, FormService formService) {
         this.expressionManager = expressionManager;
         this.protectedVarContext = protectedVarContext;
+        this.formService = formService;
     }
 
     @Override
     public ConcordExecutionContext create(Variables source) {
-        return new ConcordExecutionContext(this, expressionManager, source, protectedVarContext);
+        return new ConcordExecutionContext(this, expressionManager, source, protectedVarContext, formService);
     }
 
     @Override
     public ConcordExecutionContext create(Variables source, String processDefinitionId, String elementId) {
-        return new ConcordExecutionContext(this, expressionManager, source, processDefinitionId, elementId, protectedVarContext);
+        return new ConcordExecutionContext(this, expressionManager, source, processDefinitionId, elementId, protectedVarContext, formService);
     }
 
     @Override
@@ -61,19 +68,25 @@ public class ConcordExecutionContextFactory implements ExecutionContextFactory<C
         private static final String PROTECTED_VAR_KEY = "__protected_vars";
 
         private final ProtectedVarContext protectedVarContext;
+        private final FormService formService;
 
         public ConcordExecutionContext(ExecutionContextFactory<? extends ExecutionContext> ctxFactory,
                                        ExpressionManager expressionManager, Variables source,
-                                       ProtectedVarContext protectedVarContext) {
-            this(ctxFactory, expressionManager, source, null, null, protectedVarContext);
+                                       ProtectedVarContext protectedVarContext,
+                                       FormService formService) {
+
+            this(ctxFactory, expressionManager, source, null, null, protectedVarContext, formService);
         }
 
         public ConcordExecutionContext(ExecutionContextFactory<? extends ExecutionContext> ctxFactory,
                                        ExpressionManager expressionManager, Variables source,
                                        String processDefinitionId, String elementId,
-                                       ProtectedVarContext protectedVarContext) {
+                                       ProtectedVarContext protectedVarContext,
+                                       FormService formService) {
+
             super(ctxFactory, expressionManager, source, processDefinitionId, elementId);
             this.protectedVarContext = protectedVarContext;
+            this.formService = formService;
         }
 
         @Override
@@ -132,6 +145,44 @@ public class ConcordExecutionContextFactory implements ExecutionContextFactory<C
             super.removeVariable(key);
         }
 
+        @Override
+        @SuppressWarnings("unchecked")
+        public void form(String formName, Map<String, Object> formOptions) {
+            Map<String, Object> env = getVariables();
+
+            Map<String, Object> interpolatedOptions = (Map<String, Object>) interpolate(formOptions);
+            FormDefinition fd = new FormDefinition(formName, toFormFields((List<Object>) interpolatedOptions.get("fields")));
+
+            String txId = (String)getVariable(Constants.Context.TX_ID_KEY);
+            UUID fId = UUID.randomUUID();
+            String eventName = UUID.randomUUID().toString();
+            try {
+                formService.create(txId, fId, eventName, fd, interpolatedOptions, env);
+            } catch (ExecutionException e) {
+                throw new RuntimeException("create form error: " + e.getMessage());
+            }
+
+            suspend(eventName);
+        }
+
+        @SuppressWarnings("unchecked")
+        private List<FormField> toFormFields(List<Object> input) {
+            List<FormField> result = new ArrayList<>(input.size());
+
+            for (Object v : input) {
+                if (v instanceof FormField) {
+                    result.add((FormField) v);
+                } else if (v instanceof Map) {
+                    FormField f = formService.toFormField((Map<String, Object>) v);
+                    result.add(f);
+                } else {
+                    throw new IllegalArgumentException("Expected either a FormField instance or a map of values, got: " + v);
+                }
+            }
+
+            return result;
+        }
+
         @SuppressWarnings("unchecked")
         private Set<String> getProtectedVariableNames() {
             Set<String> result = (Set<String>) getVariable(PROTECTED_VAR_KEY);
@@ -150,12 +201,16 @@ public class ConcordExecutionContextFactory implements ExecutionContextFactory<C
 
     public static class MapBackedExecutionContext extends DefaultExecutionContextFactory.MapBackedExecutionContext implements Context {
 
+        private final ExecutionContext delegate;
+
         public MapBackedExecutionContext(ExecutionContextFactory<? extends ExecutionContext> executionContextFactory,
                                          ExpressionManager exprManager,
                                          ExecutionContext delegate,
                                          Map<Object, Object> overrides) {
 
             super(executionContextFactory, exprManager, delegate, overrides);
+
+            this.delegate = delegate;
         }
 
         @Override
@@ -166,6 +221,11 @@ public class ConcordExecutionContextFactory implements ExecutionContextFactory<C
         @Override
         public Object getProtectedVariable(String key) {
             return super.getVariable(key);
+        }
+
+        @Override
+        public void form(String formName, Map<String, Object> formOptions) {
+            ((Context)delegate).form(formName, formOptions);
         }
     }
 }