+++ sdk/src/main/java/com/walmartlabs/concord/sdk/EventService.java
@@ -23,6 +23,9 @@ package com.walmartlabs.concord.sdk;
 import java.io.Serializable;
 import java.util.Date;
 
+/**
+ * Reserved for the future use.
+ */
 public interface EventService {
 
     void onEvent(String instanceId, Date date, EventType type, Serializable data) throws ClientException;