+++ sdk/src/main/java/com/walmartlabs/concord/sdk/Task.java
@@ -20,9 +20,8 @@ package com.walmartlabs.concord.sdk;
  * =====
  */
 
-
 /**
- * Marker interface for Concord tasks.
+ * Default interface for Concord tasks.
  */
 public interface Task {